Reworking file organization to better separate into packages.
