+++ data_handling/RawDataFetcher.py
@@ -1,7 +1,7 @@
 from datetime import time, date, datetime, timedelta
 
 from utils.Logger import Logger
-from utils.date_helper import get_prev_17lands_update_time, utc_today
+from data_handling.utils.date_helper import get_prev_17lands_update_time, utc_today
 
 from data_handling.JSONHandler import JSONHandler
 from game_metadata.FormatMetadata import FormatMetadata
